+++ concordium-consensus/src/Concordium/Scheduler/ProtocolLevelTokens/Kernel.hs
@@ -18,10 +18,11 @@ class PLTKernelQuery m where
     type PLTAccount m
     getTokenState :: TokenStateKey -> m (Maybe TokenStateValue)
     getAccount :: AccountAddress -> m (Maybe (PLTAccount m))
+    getAccountIndex :: PLTAccount m -> m AccountIndex
+    getAccountByIndex :: AccountIndex -> m (Maybe (PLTAccount m))
     getAccountBalance :: PLTAccount m -> m TokenRawAmount
     getAccountState :: PLTAccount m -> TokenStateKey -> m (Maybe TokenStateValue)
     getAccountCanonicalAddress :: PLTAccount m -> m AccountAddress
-    getGovernanceAccount :: m (PLTAccount m)
     getCirculatingSupply :: m TokenRawAmount
     getDecimals :: m Word8